New localizations for the upcoming BiblioCraft 1.7 update
I believe these are to be all the translations for the next big
BiblioCraft update. I will post on my website soon with some showcase
images and a "call to translators" to help get as many languages up to
date as we can while I finish up everything else for the release.