Update
Mostly French update and small fix from the google drive.
I also had an idea. I could easily translate the missing lines for any
language usin google translate, and a chat message could be displayed on
load. The message would be "include google translation, please visit
bibliocraft.com to help improve the translation" or something like that.
As it would be a line in the file, it's easy to display only for
languages that require review.
Also want to let you know that Terrafirmacraft was just released this
week for 1.7.10