updated en_US.lang for upcoming BiblioCraft 2.0
This initially just copies over the previous localizations and I added all the changed to the top of the list. There is some cleanup that could be done to remove the old stuff that was replaced. I tried to keep most of the localizations basically the same but some changed due to changes in how I created and register certain blocks.